Add ability to store user token outside the config
Some people, myself included, like to store their configuration in Git
and make it public so that they can share tips and tricks regarding
their configuration. This is not conducive to storing security tokens.
This change adds the ability to store the Gist It security token in
a separate file `~/.atom/gist-it.token`.